+++ docs/sync.js
@@ -220,11 +220,10 @@ class Sync {
     static setLogLevel(log_level) {}
 
     /**
-     * Capture the sync log.
-     * @param {Realm.Sync~LogLevel} log_level - The log level.
-     * @param {string} message - The log message.
+     * Capture the sync client's log.
+     * @param {function(log_level, message)} logCallback - The log callback.
      */
-    static setSyncLogger(log_level, message) {}
+    static setLogger(logCallback) {}
 
     /**
      * Set the application part of the User-Agent string that will be sent to the Realm Object Server when a session