+++ CHANGELOG.md
@@ -1,10 +1,10 @@
 x.x.x Release notes (yyyy-MM-dd)
 =============================================================
 ### Enhancements
-* None.
+* Add `Realm.Sync.setLogger()` to capture the sync client's log. ([#2125](https://github.com/realm/realm-js/issues/2125) and [realm-js-private#517](https://github.com/realm/realm-js-private/issues/517))
 
 ### Fixed
-* Add `Realm.Sync.setSyncLogger()` to TypeScript definitions and API documentation. ([#2125](https://github.com/realm/realm-js/issues/2125) and [realm-js-private#517](https://github.com/realm/realm-js-private/issues/517), since v2.0.0)
+* None.
 
 ### Compatibility
 * Realm Object Server: 3.11.0 or later.